+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 convention.
+# [0.0.2](2019-09-30)
+
+
+### Bug Fixes:
+
+---
+
+### Features:
+
+* **core:** implemented getExistingUpiApps() to get list of existing upi apps
+
+
 # [0.0.1](2019-09-27)
 
 
@@ -8,4 +20,4 @@ convention.
 
 ### Features:
 
-* **core:** implement core functionality
+* **core:** implemented core functionality